🐛 Fix incorrect usage delay calculations

+++ include/gameplay/items/Weapon.hpp
@@ -6,7 +6,7 @@ class Weapon : public Consumable
 {
     public:
     Weapon() = default;
-    Weapon(const std::int32_t& damage, const std::uint32_t& magazine_capacity, const std::uint32_t& total_ammo, const float& bullet_velocity, const float& use_delay);
+    Weapon(const std::int32_t& damage, const std::uint32_t& magazine_capacity, const std::uint32_t& total_ammo, const float& bullet_velocity, const double& use_delay);
     virtual ~Weapon() = default;
 
     void use() override;
